Creating post request
Created an initial file for the post request to HCP. Also removed consoles for the list integrity after get. As nice as a visual representation would be it isn't in the scope of the task so I've moved my focus away as I already know the data is coming in correctly.